server.py: Add UNKNOWN lab and other edge cases
Add UNKOWN labs to handle unknown clients.
Other parts of the code cleaned properly.

One edge case is not done: If a mac address in a lab
is assigned an IP, and if another request is made by
the machine with same MAC addr, then another IP should
not be assigned?